migrations/2022_03_25_205530_create_table_admin_menu.php, migrations/2022_04_10_193630_create_table_admin_settings.php

+++ migrations/2022_03_25_205530_create_table_admin_menu.php
@@ -15,25 +15,26 @@ class CreateTableAdminMenu extends Migration
      */
     public function up()
     {
-        if (!Schema::hasTable('admin_menu')) {
-
-            Schema::create('admin_menu', static function (Blueprint $table) {
-                $table->bigIncrements('id');
-                $table->string('name');
-                $table->string('icon');
-                $table->string('route');
-                $table->string('action')->nullable();
-                $table->enum('type', ['item', 'resource', 'group'])->default('item');
-                $table->text('except')->nullable();
-                $table->integer('order')->default(0);
-                $table->boolean('active')->default(1);
-                $table->foreignId('parent_id')
-                    ->nullable()
-                    ->constrained('admin_menu')
-                    ->nullOnDelete()->cascadeOnUpdate();
-                $table->timestamps();
-                $table->softDeletes();
-            });
+        if (config('admin.functional.menu')) {
+            if (!Schema::hasTable('admin_menu')) {
+                Schema::create('admin_menu', static function (Blueprint $table) {
+                    $table->bigIncrements('id');
+                    $table->string('name');
+                    $table->string('icon');
+                    $table->string('route');
+                    $table->string('action')->nullable();
+                    $table->enum('type', ['item', 'resource', 'group'])->default('item');
+                    $table->text('except')->nullable();
+                    $table->integer('order')->default(0);
+                    $table->boolean('active')->default(1);
+                    $table->foreignId('parent_id')
+                        ->nullable()
+                        ->constrained('admin_menu')
+                        ->nullOnDelete()->cascadeOnUpdate();
+                    $table->timestamps();
+                    $table->softDeletes();
+                });
+            }
         }
     }
 
@@ -44,6 +45,8 @@ class CreateTableAdminMenu extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('admin_menu');
+        if (config('admin.functional.menu')) {
+            Schema::dropIfExists('admin_menu');
+        }
     }
 }
